chore: rename crypto service file to follow convention
