+++ src/externals/bridge/bridge.service.ts
@@ -3,7 +3,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { FileAttributes } from 'src/modules/file/file.domain';
 import { User } from 'src/modules/user/user.domain';
-import { CryptoService } from '../crypto/crypto';
+import { CryptoService } from '../crypto/crypto.service';
 import { HttpClient } from '../http/http.service';
 
 @Injectable()
