+++ src/externals/crypto/crypto.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import configuration from '../../config/configuration';
-import { CryptoService } from './crypto';
+import { CryptoService } from './crypto.service';
 
 @Module({
   imports: [
